Reviewing feats and fixing any prerequisite issues. Validated the crane style tree.
I think the prerequisites need to be tweaked though. RAR give a set of prereqs OR monk of a certain level. This way, a monk doesn't have to take Crane Style to gain Crane Wing. The problem with this is the description of Crane Wing. It begins "Once per round while using Crane Style,..." but the monk doesn't have to take Crane Style.  To me, that invalidates the Crane Wing ability, since it relies specifically on the use of Crane Style.  Crane Riposte is partially this way. The second half of it begins "Whenever you use Crane Wing...". So, to me, not having Crane Wing means half this feat is invalid.  To me, the Crane xxx feat prerequisite should persist regardless if the character is a monk or not.
Issue#: NEWSOURCE-130


git-svn-id: https://pcgen.svn.sourceforge.net/svnroot/pcgen/Trunk/content/notfordistribution@17122 b6a209e6-650d-0410-972b-dadffa731f64